Changed typescript definitions to cover .jsx files
