+++ Button/index.d.ts
@@ -1,19 +1,3 @@
-import MaterialComponent from '../MaterialComponent';
-import {VNode} from 'preact';
+import Button from './Button';
 
-declare interface IButtonProps {
-  dense?: boolean;
-  raised?: boolean;
-  disabled?: boolean;
-  unelevated?: boolean;
-  outlined?: boolean;
-  primary?: boolean;
-  secondary?: boolean;
-}
-export default class Button<
-  PropsType = JSX.HTMLAttributes,
-  StateType = {}
-> extends MaterialComponent<PropsType & IButtonProps, StateType> {
-  static Icon: typeof Icon;
-}
-declare class Icon extends MaterialComponent<JSX.HTMLAttributes, {}> {}
+export default Button;
