+++ Fab/index.d.ts
@@ -1,14 +1,5 @@
-import MaterialComponent from '../MaterialComponent';
-import {VNode} from 'preact';
+import Fab from './Fab';
 
-declare interface IFabProps extends JSX.HTMLAttributes {
-  mini?: boolean;
-  exited?: boolean;
-  primary?: boolean;
-  secondary?: boolean;
-}
+export default Fab;
+export * from "./Fab";
 
-export default class Fab extends MaterialComponent<IFabProps, {}> {
-  static Icon: typeof Icon;
-}
-export class Icon extends MaterialComponent<JSX.HTMLAttributes, {}> {}
