+++ FormField/FormField.d.ts
@@ -1,0 +1,8 @@
+import MaterialComponent from '../MaterialComponent';
+import {VNode} from 'preact';
+
+declare interface IFormfieldProps extends JSX.HTMLAttributes {
+  'align-end'?: boolean;
+}
+
+export default class Formfield extends MaterialComponent<IFormfieldProps, {}> {}
