+++ Switch/Switch.d.ts
@@ -1,0 +1,4 @@
+import MaterialComponent from '../MaterialComponent';
+import {VNode} from 'preact';
+
+export default class Switch extends MaterialComponent<JSX.HTMLAttributes, {}> {}
